requirements.txt
beautifulsoup4==4.6.3
numpy==1.15.2
PyInstaller==3.4
requests==2.20.0
scipy==1.1.0
youtube-dl==2018.10.5
PyQt5==5.11.3
SQLAlchemy==1.2.12
python-decouple==3.1
alembic==1.0.1